+++ commands/dns/record/record.go
@@ -28,7 +28,7 @@ func RecordCommand() *core.Command {
 	cmd := &core.Command{
 		Command: &cobra.Command{
 			Use:              "record",
-			Short:            "The sub-commands of `ionosctl dns record` allow you to manage DNS records. Records allow directing traffic for a domain to its correct location.",
+			Short:            "The sub-commands of 'ionosctl dns record' allow you to manage DNS records. Records allow directing traffic for a domain to its correct location.",
 			Aliases:          []string{"r"},
 			TraverseChildren: true,
 		},